mysite/requirements.txt
Django==1.8.5
#MySQL-Connector-Python==2.0.2
MySQL-python==1.2.3

django-autocomplete-light==2.2.6
django-multiselectfield==0.1.2
django-salmonella==0.6.1
django-suit==0.2.15
#django-suit-ckeditor==0.0.2
simplejson==3.6.3
#suit==2.0.2
wsgiref==0.1.2
django-adminplus==0.3